@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { FormGroup, FormControl, Validators } from '@angular/forms';
+import { UntypedFormGroup, UntypedFormControl, Validators } from '@angular/forms';
 
 import swal from 'sweetalert';
 import { UsuarioService } from '../services/service.index';
@@ -15,13 +15,13 @@ declare function init_plugins();
 })
 export class RegisterComponent implements OnInit {
 
-  forma: FormGroup;
+  forma: UntypedFormGroup;
 
   constructor( public _usuarioService: UsuarioService,
               public router: Router) { }
 
   sonIguales ( campo1: string, campo2: string) {
-    return ( group: FormGroup) => {
+    return ( group: UntypedFormGroup) => {
 
       let pass1 = group.controls[campo1].value;
       let pass2 = group.controls[campo2].value;
@@ -39,12 +39,12 @@ export class RegisterComponent implements OnInit {
   ngOnInit() {
     init_plugins();
 
-    this.forma = new FormGroup({
-      nombre: new FormControl(null, Validators.required ),
-      correo: new FormControl(null, [Validators.required, Validators.email]),
-      password: new FormControl(null, Validators.required),
-      password2: new FormControl(null, Validators.required),
-      condiciones: new FormControl(false)
+    this.forma = new UntypedFormGroup({
+      nombre: new UntypedFormControl(null, Validators.required ),
+      correo: new UntypedFormControl(null, [Validators.required, Validators.email]),
+      password: new UntypedFormControl(null, Validators.required),
+      password2: new UntypedFormControl(null, Validators.required),
+      condiciones: new UntypedFormControl(false)
     }, { validators: this.sonIguales('password', 'password2') });
 
     this.forma.setValue({
